use unsigned long not int for cpu% calcs. fixes bug# 1351686.  patch bu boojit
git-svn-id: https://conky.svn.sourceforge.net/svnroot/conky/trunk/conky@375 7f574dfc-610e-0410-a909-a81674777703

+++ src/conky.h
@@ -340,11 +340,11 @@ struct process {
 	pid_t pid;
 	char *name;
 	float amount;
-	unsigned int user_time;
-	unsigned int total;
-	unsigned int kernel_time;
-	unsigned int previous_user_time;
-	unsigned int previous_kernel_time;
+	unsigned long user_time;
+	unsigned long total;
+	unsigned long kernel_time;
+	unsigned long previous_user_time;
+	unsigned long previous_kernel_time;
 	unsigned int vsize;
 	unsigned int rss;
 	unsigned int time_stamp;
